Adds settings and updates info view
Implements a settings view allowing users to toggle sound effects and dark mode, and provides links to the terms of service and help & support.

Updates the info/help view to include Telegram links for the developers.
Replaces the gear icon button on the main view with the settings view.
Also, update some texts.